pyproject.toml
[project]
name = "horus"
version = "0.1.0"
description = ""
authors = [
    {name = "pkalsi97",email = "pranjulkalsi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv (>=0.9.9,<0.10.0)",
    "google-genai (>=1.9.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
